Remove js/data.js from management by git
The data.js file presents data as js data structures. The problem is that
the data changes rapidly, so coupling the data refreshes to code
deployment makes for a poor workflow.  The data.js file will continue
to exist outside of git, and will be deployed (most likely) via rsync.